Conflicts:
	identity/identity-core/src/main/groovy/com/junbo/identity/core/service/validator/impl/UserTFAAttemptValidatorImpl.groovy

+++ crypto/crypto-data/src/test/groovy/com/junbo/crypto/data/RepositoryTest.groovy
@@ -43,7 +43,7 @@ class RepositoryTest extends AbstractTestNGSpringContextTests {
         SecureRandom random = new SecureRandom()
         MasterKey masterKey = new MasterKey()
         masterKey.setEncryptValue(UUID.randomUUID().toString())
-        masterKey.setKeyVersion(Math.abs(random.nextInt()))
+        masterKey.setKeyVersion(Math.abs(random.nextLong()))
         masterKey.setCreatedBy(123L)
         masterKey.setCreatedTime(new Date())
         MasterKey newMaster = masterKeyRepo.create(masterKey).get()
